+++ bot.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -100,9 +101,24 @@ func New(token string, options ...Option) (*Bot, error) {
 	return b, nil
 }
 
+// ID returns the bot ID
+func (b *Bot) ID() int64 {
+	id, err := strconv.ParseInt(strings.Split(b.token, ":")[0], 10, 64)
+	if err != nil {
+		return 0
+	}
+
+	return id
+}
+
 // SetToken sets the bot token
 func (b *Bot) SetToken(token string) {
 	b.token = token
+}
+
+// Token returns the bot token
+func (b *Bot) Token() string {
+	return b.token
 }
 
 // StartWebhook starts the Bot with webhook mode
@@ -148,11 +164,13 @@ func (b *Bot) error(format string, args ...interface{}) {
 	b.errorsHandler(fmt.Errorf(format, args...))
 }
 
+// True and False returns the pointer to bool
 func True() *bool {
 	b := true
 	return &b
 }
 
+// False and True returns the pointer to bool
 func False() *bool {
 	b := false
 	return &b